moved input trigger related startup code to start_input_triggers.py
